solved problem with api
