+++ index.html
@@ -13,60 +13,60 @@
     <div class="messenger-box">
         <div class="contact-list">
             <div class="contact" id="person1">
-                <div class="name">Jan Pierwszy</div>
-                <div class="status">Online</div>
+                <div class="name"></div>
+                <div class="status"></div>
                 <div class="status-dot"></div>
             </div>
             <div class="contact" id="person2">
-                <div class="name">Jan Drugi</div>
-                <div class="status">Online</div>
+                <div class="name"></div>
+                <div class="status"></div>
                 <div class="status-dot"></div>
             </div>
             <div class="contact" id="person3">
-                <div class="name">Jan Trzeci</div>
-                <div class="status">Online</div>
+                <div class="name"></div>
+                <div class="status"></div>
                 <div class="status-dot"></div>
             </div>
             <div class="contact" id="person4">
-                <div class="name">Jan Czwarty</div>
-                <div class="status">Online</div>
+                <div class="name"></div>
+                <div class="status"></div>
                 <div class="status-dot"></div>
 
 
             </div>
             <div class="contact" id="person5">
-                <div class="name">Jan Piąty</div>
-                <div class="status">Online</div>
+                <div class="name"></div>
+                <div class="status"></div>
                 <div class="status-dot"></div>
 
             </div>
             <div class="contact" id="person6">
-                <div class="name">Jan Szósty</div>
-                <div class="status">Online</div>
+                <div class="name"></div>
+                <div class="status"></div>
                 <div class="status-dot"></div>
 
             </div>
             <div class="contact" id="person7">
-                <div class="name">Jan Siódmy</div>
-                <div class="status">Online</div>
+                <div class="name"></div>
+                <div class="status"></div>
                 <div class="status-dot"></div>
 
             </div>
             <div class="contact" id="person8">
-                <div class="name">Jan Ósmy</div>
-                <div class="status">Online</div>
+                <div class="name"></div>
+                <div class="status"></div>
                 <div class="status-dot"></div>
 
             </div>
             <div class="contact" id="person9">
-                <div class="name">Jan Dziewiąty</div>
-                <div class="status">Online</div>
+                <div class="name"></div>
+                <div class="status"></div>
                 <div class="status-dot"></div>
 
             </div>
             <div class="contact" id="person10">
-                <div class="name">Jan Dziesiąty</div>
-                <div class="status">Online</div>
+                <div class="name"></div>
+                <div class="status"></div>
                 <div class="status-dot"></div>
 
             </div>
@@ -83,7 +83,7 @@
             <input type="text" placeholder="Text here..." />
         </form>
     </div>
-    <script src="./dist/main.js"></script>
+    <script async src="./dist/main.js"></script>
 </body>
 
 </html>
